Add Docker Support (#2)

+++ scripts/fpp_uninstall.sh
@@ -2,8 +2,12 @@
 
 echo "Running pulsemesh-connector uninstall script"
 
-sudo systemctl --now disable pulsemesh-connector.service
-sudo apt-get purge -y pulsemesh-connector
-sudo rm /etc/apt/sources.list.d/pulsemsh.list
+if [ -f "/.dockerenv" ]; then
+    echo "Running in docker, skipping..."
+else
+    sudo systemctl --now disable pulsemesh-connector.service
+    sudo apt-get purge -y pulsemesh-connector
+    sudo rm /etc/apt/sources.list.d/pulsemsh.list
 
-echo "Removed pulsemesh-connector"
+    echo "Removed pulsemesh-connector"
+fi
